+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # DJS Level Importer changelog
 
+## 2.0.0 (unreleased)
+
+- Breaking changes:
+  - Reworked a bit the data return types. Breaking. Now `FullUserLevels` is `StandardUserLevels` (and properties `current_lvl_xp` and `next_lvl_xp` are gone), with `FullUserLevels` becoming a new interface with `rewards` property (not yet in use).
+- Changed:
+  - Now the package is zero-dependency while still supporting all bots. Install ~75% smaller.
+
 ## 1.2.0
 
 - Added bots: